Added all the code and classes.

+++ CS-421/ExpressionTester/src/DoubleValue.java
@@ -2,7 +2,7 @@
 public class DoubleValue implements Number {
 	private double value;
 	
-	public DoubleValue(int value) {
+	public DoubleValue(double value) {
 		this.value = value;
 	}
